+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Raise when trying to cache a belong_to association with a scope. Previously the scope was ignored on a cache hit (#323)
 - Remove deprecated `never_set_inverse_association` option (#319)
 - Fix column name in the preload association query when using custom primary keys (#338)
+- Fix inverse asoociation setting for cache_has_many without `:embed => true`
 
 #### 0.5.1